RUM-11509: increasing minSdk for SDK
