+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/recorder/ViewUtilsInternal.kt
@@ -8,7 +8,6 @@ package com.datadog.android.sessionreplay.internal.recorder
 
 import android.annotation.SuppressLint
 import android.graphics.drawable.Drawable
-import android.os.Build
 import android.view.Display
 import android.view.View
 import android.view.ViewStub
@@ -35,13 +34,9 @@ internal class ViewUtilsInternal {
 
     internal fun isOnSecondaryDisplay(view: View): Boolean {
         val display = view.display
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            display != null &&
-                display.displayId != Display.DEFAULT_DISPLAY &&
-                display.displayId != Display.INVALID_DISPLAY
-        } else {
-            display != null && display.displayId != Display.DEFAULT_DISPLAY
-        }
+        return display != null &&
+            display.displayId != Display.DEFAULT_DISPLAY &&
+            display.displayId != Display.INVALID_DISPLAY
     }
 
     @Suppress("UnsafeThirdPartyFunctionCall") // NPE cannot happen here
