+++ sample/tv/src/main/java/com/datadog/android/tv/sample/EpisodeRecyclerView.kt
@@ -7,7 +7,6 @@
 package com.datadog.android.tv.sample
 
 import android.annotation.SuppressLint
-import android.os.Build
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -71,12 +70,7 @@ internal class EpisodeRecyclerView private constructor() {
 
         private fun updateEpisodeHighlight(selected: Boolean) {
             val colorRes = if (selected) R.color.dd_purple_200 else R.color.text_default
-            val textColor = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                textView.resources.getColor(colorRes, null)
-            } else {
-                @Suppress("DEPRECATION")
-                textView.resources.getColor(colorRes)
-            }
+            val textColor = textView.resources.getColor(colorRes, null)
             textView.setTextColor(textColor)
         }
